gfdsa

+++ BE/Configuration.cs
@@ -10,7 +10,7 @@ namespace BE
         public static int GuestRequestKey;
         public static int HostKey;
         public static int OrderKey=0;
-        public static double Commission = 10; //commission for orders
+        public static int Commission = 10; //commission for orders
         public static int Password = 999;
         public static string MangerPassword = "9999";
         static public String getMangerPassword() { return MangerPassword; }
